requirements.txt
Flask==0.10.1
requests==1.2.3
python-firebase==1.2
